Tribal Signal Fire
This mod adds two new buildings to the game; the 'Signal Fire' and the 'Tribal Trade Beacon'. Tribal colonies were never able to communicate with other factions in the game so this mod will allow you to do just that. The SignalFire is built, looks and requires to be refuled just like the campfire but pawns can interact with it as though it is a comms console. You cannot communicate with orbital trade ships, just ground factions.
     The Tribal Trade Beacon acts just like the Orbital Trade Beacon but does not require power and is built with wood.